EDU-795 Extract functionality into utils modules
